`InAppBrowser`'s `API` is technically public in terms of calling `window["functionName"]( ...data )`. While all it can do is close or open a link - that is all it can do Despite this `API` now requires a certain `key` to be passed in to `window["functionName"]( ...data )` function so that using `Developer Tools` it cannot be envoked as whoever envokes it will have to require the internal key just for using the function. While this measure is not perfect, it still prevents the envoking of `InAppBrowser`'s `API` from `Developer Tools`.
`InAppBrowser`'s `API` is technically public in terms of calling `window["functionName"]( ...data )`. While all it can do is close or open a link - that is all it can do
Despite this `API` now requires a certain `key` to be passed in to `window["functionName"]( ...data )` function so that using `Developer Tools` it cannot be envoked as whoever envokes it will have to require the internal key just for using the function.
While this measure is not perfect, it still prevents the envoking of `InAppBrowser`'s `API` from `Developer Tools`.