Created internal `API` for internal use. The main idea is to expose `React Hooks API` of components centrally to other components that may not be linked in any other way. With the use of internal `API` - Indvidual components' `API` is no longer relient on `window` object. This also means that it is no longer exposed using `Developer Tools` at least not easily. Furthermore the concept of validating for the `key` when using `window["function"]` is no longer required or neccessary.
Created internal `API` for internal use.
The main idea is to expose `React Hooks API` of components centrally to other components that may not be linked in any other way.
With the use of internal `API` - Indvidual components' `API` is no longer relient on `window` object.
This also means that it is no longer exposed using `Developer Tools` at least not easily. Furthermore the concept of validating for the `key` when using `window["function"]` is no longer required or neccessary.