Added a custom ` React Hook` - still do not fully understand it yet. It is called `UseFunctionAsState`. This allows setting functions for dynamic clicking of button. Usually `React.useState( () => { ... }` would work and it does work but not in the way I expect. This way function will run immediatley as the function is set however `React.useState( () => { ... }` returns the value of function not returning the actual function itself. Custom hooks are meant for this sort of thing. I use it and it work but not quite fully understand the depth of technicality.
Added a custom ` React Hook` - still do not fully understand it yet. It is called `UseFunctionAsState`.
This allows setting functions for dynamic clicking of button. Usually `React.useState( () => { ... }` would work and it does work but not in the way I expect. This way function will run immediatley as the function is set however `React.useState( () => { ... }` returns the value of function not returning the actual function itself.
Custom hooks are meant for this sort of thing. I use it and it work but not quite fully understand the depth of technicality.