Added a `Side Projects` screen. Added relevant `CMS` for it. Updated `Routes` accordingly. Fixed a bug for `Linear` progress bar component where it would hide very after `rendering` because there is no immediate need for it to be shown. This is a design flaw or not the right way. At this point I have not yet found a "right" approach hencing fall for what works as last resort.
Added a `Side Projects` screen.
Added relevant `CMS` for it.
Updated `Routes` accordingly.
Fixed a bug for `Linear` progress bar component where it would hide very after `rendering` because there is no immediate need for it to be shown. This is a design flaw or not the right way. At this point I have not yet found a "right" approach hencing fall for what works as last resort.