`getCardActions` of `Utils` now offically supports `title` as oppose to relying on the `text` property.  `title` is directly injected into `getCardActions` which is whatever the `Card`'s title is. This does mean that if a particular card has more than one `InAppBrowser` instances on a click - all of them will have the same `title` which is fine because it is all part of that `Card`'s title.
`getCardActions` of `Utils` now offically supports `title` as oppose to relying on the `text` property. 
`title` is directly injected into `getCardActions` which is whatever the `Card`'s title is.
This does mean that if a particular card has more than one `InAppBrowser` instances on a click - all of them will have the same `title` which is fine because it is all part of that `Card`'s title.