Finally managed to virtualise `SimpleCardList` using
Finally managed to virtualise `SimpleCardList` using `react-tiny-virtual-list` the way how I wanted to so that it appends the new content underneath the current content and not doing too much rendering without any duplication of initial set of data. The rerendering still happens but in managable way and is only noticable due to image being reloaded again. Overall it is a good effort.