Update README.md
The contents of `Dashboard` is now fully `CMS`ified for easy updating of content without changing the the logic aspect.
`Dashboard` content is quite different and unique than before so it made sense to design it first with hand than putting it first in `CMS`. 
Had I first tried to put it in `CMS` - I would probably still be trying to figure out a way to do it. Early optimization is the root of all evil.